Compare mime types directly in HttpTranslator
If the contentType was created via ContentType.parse the
charset of the translated request would always be modified
to be ISO_8859_1 even if the content type was application/json.

Tests to validate charset returned from HttpTranslator.getHttpEntity

HttpEntities returned from HttpTranslator.getHttpEntity should always
have a
charset of ISO_8859_1 unless the contentType is APPLICATION_JSON, in
which case
it should stay UTF-8.

Cleanup commented code in HttpTranslator

Signed-off-by: Paul LeMarquand <plemarquand@qnx.com>