Fix 475112: Only prefer server certificate type RawPublicKey if no trust store has been configured.
This way a DtlsConnector can be configured to either request
RawPublicKeys or X.509 certificates from server based on user
preferences. Note that we currently do not provide any way to validate a
RawPublicKey received from a server, e.g. by comparing a hash. We only
make sure that the server is in possession of the matching private key.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>