Fix 477074. The encoding of a Raw Public Key contained in a CERTIFICATE message now complies with RFC 7250.
In particular the RPK's public subject info data is prepended with only
three bytes indicating the length of the public subject info byte array.

Note that this fix probably breaks interoperability with released
versions of TinyDTLS which also do not comply with RFC 7250. However, a
fix has been committed to TinyDTLS on 2015-08-26, so any subsequent
builds should again be compatible with Scandium.