Increase inbound DatagramPacket buffer size to theoretical maximum.
The size of the byte array used for reading from the network socket was
set to the Max. Payload Size configuration value so far. This works for
sufficiently large values (e.g. 4096 bytes) because in the existing test
cases no message of that size was ever sent. However, we cannot make any
reasonable assumption regarding the size of datagrams sent by "real"
clients. Thus, it seems reasonable to expect the "worst", i.e. the
maximum size allowed by UDP.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>