Use SessionListener to trigger sending of pending APPLICATION messages.
Sending an APPLICATION message to a peer for which no DTLS session
exists yet triggers a new handshake with the peer. Previously it was the
responsibility of the Handshaker class to send the APPLICATION layer
message to the peer once the handshake has been completed.

Handshakers signal the establishment of a session by means of the
SessionLifecycle's sessionEstablished() callback method. In order to
improve separation of concerns pending APPLICATION layer messages are
now sent by the DTLSConnector on this callback.