Add support for session fail over between nodes.
Introduced a SessionCache strategy interface for storing and retrieving
state of an established DTLS session to and from e.g. a persistent
store.
The session information can be shared by multiple nodes and e.g. be used
to resume a session with a client in case the original node that the
client has established the session with, fails.