Extend OptionNumberRegister test to check ordering
If OptionNumberRegister values are registered in ascending order,
an efficient EnumMap can be used to check options in an OptionSet
and this will allow to refactor OptionSet class so no need of
sorting and ordering options by protocol value number to comply
with RFC-7252.  EnumMap allows to traverse the container in
OptionNumberRegister natural order, so no need to make a list, and
sort it before message serialization.

Signed-off-by: Luis Colorado <lcu@basen.net>