ClientAsynchronousTest rewritten
ClientAsynchronousTest does all assertions in the callback routine, that
doesn't get called in case of absent server.  I have tried to run it
from eclipse, and got completely surprised on seeing it pass, when eclipse was
still asking for permissions to open the udp listening
socket of the server.

I have rewritten this tests to only make response available and properly
wait for the response to arrive, and give a complete test timeout of 5s.
So, in case of failing server the test does not pass.

Signed-off-by: Luis Colorado <lcu@basen.net>