[bug #68] fix broken DTLS handshake
295217d introduced a regression which caused the last flight of a
successful handshake to always be retransmitted again and again. This
was due to the fact that we switched from java.util.Timer to
java.util.concurrent.ScheduledExecutorService for scheduling
retransmission of flights. The mechanism for cancelling scheduled tasks
in the latter is different from the former which had not been considered
when we made the change.