Ok, this is an almost complete stab at implementing CoAP-over-TCP RFC. Things worth noting about this PR:
Made an attempt to keep client interface fully backwards compatible. This means users of CoapServer, CoapClient and Request/Response objects should not care whether protocol runs over TCP or UDP. The only downside is that it allows client ability to send ACKs and RSTs, which get silently dropped by the TCP stack. Given that you need to opt-in into using TCP stack, I think that's a reasonable trade off.

Now to the actual implementation. Introduced interfaces for Matcher, DataSerializer, DataParser and created split implementations UDP and TCP (if anybody has better ideas for names I'm all ears). Also, introduced TcpCoapStack, which is simplified version of Udp one. Chose not to introduce an interface there.

On the connector side, implemented Tcp client and server connectors. Note, that they are asymmetrical - which I guess is expected given the nature of TCP. I don't think this violates any promises of Californium, as one already needs to instantiate CoapServer and CoapClient separately.

Went with Netty implementation, just like in Simon's branch. But used later version of Netty as it gave support for connection pooling, and tuned some parameters based on end to end tests.

Head to ThroughputClient.java and ThroughputServer.java for examples of how to use TCP connectors. One pleasant side effect of moving to the new TCP spec is getting rid of 16 bit MIDs, which made a big impact on throughput. This was a major factor for us (AWS).

Lastly, no TLS implementation at this point. If this looks like a reasonable approach should be trivial to add that.

Signed-off-by: Joe Magerramov <joe.magerramov@gmail.com>