Implementation of TCP connector for Californium per CoAP over TCP RFC.
* Uses Netty 4.1 for non-blocking IO. Implementation splits into Server
and Client TCP connector, with expectation that the client connector
does not listen on a port.
* Split Matcher into Udp and Tcp Matchers, with the latter not keeping
track of exchanges by MID.
* Split Stack into Udp and Tcp stack, with the later not having
reliability layer (but keeping blockwise, per TCP RFC recommendation)
* Changed serializers and parsers into instantiable classes to support
parsing TCP and UDP payloads.
** Refactored DataParser and DataSerializer a bit to make both symmetric
and only delegate parsing of header to individual implementations.
* Introduced Tcp Server and Client bandwidth tests that attempt to send
large amounts of requests and responses.
* Changes to EndpointManager to allow creating an endpoint with TCP
connector whenever coap+tcp scheme is used.
* With introduction of connector.isSchemeSupported() the whole endpoint
management can likely be improved, but did not want to boil the ocean.

Still todo:

    1) Add call back from the TcpConnector when a connection is lost
(per RFC all exchanges with that remote endpoint must be closed)
    2) Add ability for TcpServerConnector to open outbound connections.
This is desirable when application is reusing the same endpoint between
CoapClient and CoapServer

Signed-off-by: Joe Magerramov <joemag@amazon.com>