Fixes for https://github.com/eclipse/californium/issues/72 and https://github.com/eclipse/californium/issues/73.
Prior to this fix, if a response to the request (or in UDP stack response for acknowledged request) never arrived we'd leak Exchange in the Matcher maps. We'd also never call back client's CoapHandler.onError() potentially.

This change observes Request cancellation, and if a request is cancelled triggers Exchange.setComplete() method.

Signed-off-by: Joe Magerramov <joe.magerramov@gmail.com>