[fix #75] Server Treats NON-Message as CON-Message
Request#type is now Type.CON by default.

CoapClient: Move setting the request type to the convenience methods.
Client code can use the advanced() methods for sending a request that
it has configured manually whereas the convenience methods can be used
in the established way.

Uses Request#uri (if not blank) in the advanced() methods instead of
CoapClient#uri.

Conflicts:
	californium-core/src/main/java/org/eclipse/californium/core/CoapClient.java

Signed-off-by: Eugene Nikolaev <eug.nikolaev@gmail.com>