Support creation of and session resumption based on Session Tickets.
Session Tickets are defined by RFC 5077 to support "Session Resumption
without Server-Side State". As a first step, DTLSSession can now create
a SessionTicket containing the session's current crypto params that are
necessary to resume the session in an abbreviated handshake.

Full support for RFC 5077 remains to be implemented, i.e. we do not
(yet) support the hello extension for echanging the session ticket with
a client. However, the SessionTicket can be used to e.g. share sessions
between multiple server nodes, in particular when used as the
serialization format used in a SessionCache implementation.