Changed the visibility of observeAndWait() and observe() to public
Changed the visibility of observeAndWait(request,handler) and
observe(request,handler) to public as per discussion on the issue
https://github.com/eclipse/californium/issues/92 ..

Both the methods now checks that passed request has observer set on it,
if the observer is not set it raises IllegalArgumentException.

Test case also has been added for both methods.

Signed-off-by: Praful Bhatnagar <prafulbhatnagar@gmail.com>