Increase "cool down" time after releasing sockets.
Some test fixtures bind and unbind a socket to an address that the test
itself then needs to re-bind to. This fails sporadically on our HIPP
server.
Also reduce parallelism of execution of "medium" tests down to 1 class
at a time to prevent race conditions when multiple tests try to bind to
the same address.