[#141] Allow Serializer to be used without caching array in Message.
Add method to serialize a Request to a byte array without storing the
bytes in the Message.bytes property. This way, client code that is only
interested in getting the wire format encoding for a request does not
need to make a decision on a particular wire format to be used when the
the message is ultimately sent out.

Also improved JavaDocs and test cases.