[#159] Introduce config param for limiting blockwise buffer size.
The MAX_RESOURCE_BODY_SIZE option can be used to limit the buffer size
for doing transparent blockwise transfers or for disabling transparent
blockwise handling altogether by setting it to 0.

When doing a transparent blockwise transfer the BlockWise layer now also
uses the size2 option to indicate the overall body size being
transferred to a peer. When receiving data it uses the size1 option to
determine the buffer size required.

If transparent blockwise handling is disabled, incoming requests
containing a block1 option are forwarded to the application layer. If
the body size exceeds the configured mx buffer size, a 4.13 error is
returned.

Conflicts:
	californium-core/src/main/java/org/eclipse/californium/core/network/stack/BlockwiseLayer.java
	californium-core/src/test/java/org/eclipse/californium/core/test/lockstep/BlockwiseClientSideTest.java
	californium-core/src/test/java/org/eclipse/californium/core/test/lockstep/BlockwiseServerSideTest.java