Add TCP correlation context.
Used to distinguish different connections between the same endpoints.
The "localhost/127.0.0.1" mainly used for logging in the TCP client is
replaced by "any/0.0.0.0" because it reaches also targets on other
hosts, what would not be possible with "localhost".
The ConnectorTestUtil introduces common functions for TCP test. The
already available tests may be adpated to use this in a follow up PR.
This PR is the first part for correlation contexts for TCP based
connections. The larger part would be TLS and will follow, when this
get's merged.

Signed-off-by: Achim Kraus <achim.kraus@bosch-si.com>