[#216] Use a striped executor for processing inbound records.
Added a striped executor for scheduling the parallel processing of
records while maintaining the order of reception per peer. This way the
records can be processed using multiple threads. When using an executor
with an appropriately sized thread pool, the processing of one handshake
message does not block other ongoing handshakes.