Fix broken test depending on MTU.
The "testReceivingMessagesInBadOrderDuringHandshake" depends on the MTU,
because it expects record and not flights. Running the test on windows
set the MTU to default of 1280 and so the second flight is split into
two records which causes the test to fail.
This fix will therefore wait for the flight (using a timeout) and so
overcome the MTU dependency.

Signed-off-by: Achim Kraus <achim.kraus@bosch-si.com>