[#187] Prevent memory-leak caused by canceled but registered exchanges.
If an outbound request gets canceled before it is sent it might still
get registered with the ExchangeStore. In order to prevent a memory leak
in the ExchangeStore we therefore make sure that the Exchange gets
"completed" again if it is canceled before it is sent out via the
Connector in the Outbox.