[#187] Do not try to unregister unsent messages on exchange completion.
The Matcher asserts that the request of a completed exchange has an MID
and token before trying to unregister the exchange from the
exchange maps.

If an outbound request gets canceled before it is sent it might still
get registered with the exchange maps. In order to prevent a memory leak
we therefore make sure that the Exchange gets "completed" again if it is
canceled before it is sent out via the Connector in the Outbox.