[#511] Use Uri-Host as virtual host name in SNI extension.
When negotiation of a new DTLS session is triggered by means of sending
a CoAP request, the DTLSConnector now uses the value of the Uri-Host
contained in the request for the virtual host name to convey in the
Server Name Indication extension in its CLIENT_HELLO message.

The ServerNameResolver is therefore obsolete.