add new property to configure how respond to early blockwise negociation
Currently when a client asked for a resource with early blockwise negociation and the content fitted into a single response the block2 option wasn't included in the response though this can be confusing for some client library which expects to receive it with the "m" flag set to false.
Now by default the current process is used (no block2 option) but by setting the NetworkConfig.Keys.BLOCKWISE_STRICT_BLOCK2_OPTION to true the blockwise layer will include the block2 option even if the conte$

This fix the use case when early block2 negociation is used by the client but the response fits into a single. (issue #810)

Signed-off-by: Mickael Barbin <mickaelbarbin89@gmail.com>