Reduce scope of Serializable to AlertMessage.
Passing the peer's address into the DTLSMessages is mainly used for
logging and to generate a AlertMessage on failure for such a
DTLSMessages. Maybe in the future we can get rid of passing the peer
address around, which would then obsolete the AbstractMessage at all.
But for now, live with that workaround.

Signed-off-by: Achim Kraus <achim.kraus@bosch-si.com>