OSCORE: Improved ID Context handling and general fixes
Additions to prepare the OSCORE code for implementation of the functionality
in Appendix B.2 of the OSCORE draft text. It describes how to securely
derive new security contexts after a crash or power loss:
https://tools.ietf.org/html/draft-ietf-core-object-security-16#appendix-B.2

The solution relies heavily on using the ID Context parameter for deriving
new contexts. This commit updates the code to handle the ID Context better,
for instance parsing it properly if included in incoming messages. Further
updates are needed to finalize the functionality described in the appendix.

Furthermore some other fixes are included in this commit:

 - Updates to various comments
 - Inclusion of JUnit tests for message decryption and encryption
 - Added JUnit tests based on test vectors in draft for AAD, nonce,
   sender/recipient key and common IV
 - Reduction from 4 to 1 methods for AAD generation
 - Fix in AAD generation when a response contains a partial IV

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>