OSCORE: Initial support for method to securely re-generate contexts
This commit adds initial support for the method described in Appendix B.2.
of the OSCORE draft for securely re-generating contexts:
https://tools.ietf.org/html/draft-ietf-core-object-security-16#appendix-B.2

The idea is to perform context re-derivation by essentially using a message
exchange together with the Context ID field in the OSCORE option to securely
generate a new shared context.

This can for instance be used when one device has lost power and information
about the mutable parts of a context (e.g. sequence number) but retains
information about static parts (e.g. master secret).

This commit includes initial support for this functionality together with a
JUnit test to test it. It also includes functionality in the OSCORE code to
dynamically re-generate context information depending on the incoming Context ID.
Logic was also added to enable controlling whether messages generated from a
particular context will include Context IDs (that is not always needed and can
increase the message overhead). A further update will come to finalize the full
functionality as described in the appendix.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Fixed formatting of some if statements

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Updated two comments to improve Javadoc

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Using LOGGER instead of System.out.print to print incoming KID Context

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Added CoapNetworkRule and set @Ignore rule in ContextRederivationTest

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Removed parentheses from strings in ErrorDescriptions

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Exception when trying to include a Context ID for a context without one

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Added IllegalStateException info to Javadoc for setIncludeContextId

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>