OSCORE: Update in functionality for securely re-generating contexts
This commit updates the support for the method described in Appendix B.2.
of the OSCORE draft for securely re-generating contexts:
https://tools.ietf.org/html/draft-ietf-core-object-security-16#appendix-B.2

The idea is to perform context re-derivation by essentially using a message
exchange together with the Context ID field in the OSCORE option to securely
generate a new shared context. This can for instance be used when one device
has lost power and information about the mutable parts of a context (e.g. sequence
number) but retains information about static parts (e.g. master secret).

This commit updates support for this functionality to let the client set a
flag to indicate that it has lost the mutable parts of an OSCORE context.
When that flag is set the first message of the context re-derivation
procedure is immediately sent which will be followed by the response from
the server. After this the client can proceed with the actual request it
intends to send which will act as the third message of the context
re-derivation procedure, followed by the response from the server.

In addition the JUnit test for testing this functionality has been updated
to be able to run stand-alone without an external server application.

The end goal is that the first message of the re-derivation procedure is
not sent at the moment when the client indicates that it has lost the mutable
parts of an OSCORE context but rather not sent until the client is actually
interested in sending a request to the server. A further commit will be
created for this.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>

Small fixes from feedback to pull request

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>