OSCORE: Implements setting of peer identity and Principal for OSCORE
This commit adds a new Principal for endpoint information when using OSCORE.
It will be set in the endpoint context when communicating using OSCORE.

Currently it will only update the endpoint context if it is of type
AddressEndpointContext and if the Principal is of type OSCorePeerIdentity.
This means that it will not overwrite Principal information set by
Scandium when using DTLS. So if both DTLS and OSCORE are used DTLS
peer identity information will take precedence.

This should resolve the following issue:
https://github.com/eclipse/californium/issues/974

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>