OSCORE: Implements setting OSCORE information in endpoint contexts
This commit creates functionality that adds strings with information
about the OSCORE context used to the source or destination endpoint
context for a CoAP message. These strings will be set in the endpoint
context when communicating using OSCORE.

In certain situations an application may need to know OSCORE-related
information about a CoAP message. For instance it may need to know
which OSCORE Sender ID a message was transmitted from. The
MapBasedEndpointContext contains a map of strings that can be used to
add arbitrary information. This commit makes it so that OSCORE sets
some relevant strings in endpoint contexts when OSCORE is used. These
strings can then be used by the application to retrieve information
related to OSCORE from the endpoint context of a message.

See the new JUnit tests in this commit for a demonstration of how this
functionality can be used. Note that outgoing requests are handled in
a special way with the new EndpointContextOperator functionality. For
outgoing requests a new destination endpoint context will be created
which is available via a callback using a MessageObserver. For outgoing
responses the destination endpoint context used will be the source
endpoint context created for the incoming request.

This commit can help resolve the following issue:
https://github.com/eclipse/californium/issues/974
The previous approach of setting the PeerIdentity with a Principal for
OSCORE (see https://github.com/eclipse/californium/pull/1002) was
changed after discussion in that thread.

This is a new version of the previous PR on the same topic:
https://github.com/eclipse/californium/pull/1027

Other small changes in this commit:
- Various formatting and comment fixes
- Moved flag indicating if responses will have a Partial IV to OSCoreCtx
- Set content format for error messages to text/plain
- Removed some code related to the old singleton from HashMapCtxDB

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>