OSCORE: Added case when context re-generation is initiated by the server
Added case for the context re-generation functionality where it is
the server that loses the mutable parts of the OSCORE context and
thus takes initiative to start the context re-derivation procedure.

That is the client will not explicitly send Request 1 as described
in the context re-derivation procedure below. Rather the client
will sends a normal request and since it is the server that has lost
the mutable parts of the context it is the server that will then
continue with the re-derivation procedure.
https://tools.ietf.org/html/rfc8613#appendix-B.2

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>