OSCORE: Support for (outer) Block-Wise functionality
This commit adds further support for using OSCORE with Block-Wise requests
and responses. It implements the "outer Block-Wise" mode of OSCORE where
the Block CoAP options are not encrypted.

A proxy could split a large OSCORE protected request before forwarding it to
the server (or split a response before forwarding it to the client). In such
case the block options will be unencrypted and external. That is, the
individual block messages are not protected with OSCORE, as is the case with
inner Block-Wise, but rather the whole message (before being split) is
protected with OSCORE. Thus the message must be re-assembled by the block-wise
layer before it is unprotected by OSCORE.
See https://tools.ietf.org/html/rfc8613#section-4.1.3.4.2

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>