OSCORE: Improvements to option handling and replay window
This commit updates the handling of the OSCORE option,
including separate option encoder/decoder classes,
and now stores the entire OSCORE option in the exchange.
This avoids using a separate map to keep track of the
sequence number to use for processing incoming or
outgoing responses. It also enables improvements to
the replay window functionality.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>