OSCORE: Add OSCORE support to plugtest server and client
This commit adds support for OSCORE to the plugtest server and client.
This includes a resource on the plugtest server that can only be
accessed using OSCORE (/oscore), and a resource from which current
OSCORE configuration parameters can be retrieved (/oscoreInfo). In
addition, the README file has been expanded with key material,
information and settings for OSCORE support in the plugtest
server/client.

The OSCORE support in the plugtest server supports the Appendix B.2
context rederivation procedure, and the plugtest client will use it
by default. This is because otherwise requests from new clients would
be considered replays (as the server's replay window is filled up
from earlier clients). Clients that do not support the Appendix B.2
procedure can retrieve an appropriate Sender Sequence Number to use
and the current ID Context from the resource /oscoreInfo using
a plain CoAP request.

Signed-off-by: Rikard Höglund <rikard.hoglund@ri.se>